requirements.txt
elevenlabs==0.2.26
Flask==2.3.3
fuzzywuzzy==0.18.0
moviepy==1.0.3
numpy==1.24.4
opencv-python==4.8.0.76
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
urllib3==2.0.4
websockets==11.0.3
dtw-python
python-Levenshtein==0.21.1